(PC-28466)[PRO] refactor: remove index files in image uploader
