--- pro/src/components/ImageUploader/AppPreviewOffer/index.ts
@@ -1,1 +1,0 @@
-export { default as AppPreviewOffer } from './AppPreviewOffer'
