--- pro/src/components/ImageUploader/AppPreviewVenue/index.ts
@@ -1,1 +1,0 @@
-export { default as AppPreviewVenue } from './AppPreviewVenue'
