--- pro/src/components/ImageUploader/ButtonAppPreview/ModalAppPreview/index.ts
@@ -1,1 +1,0 @@
-export { default as ModalAppPreview } from './ModalAppPreview'
