--- pro/src/components/ImageUploader/ButtonAppPreview/index.ts
@@ -1,1 +1,0 @@
-export { default as ButtonAppPreview } from './ButtonAppPreview'
