--- pro/src/components/ImageUploader/ButtonImageDelete/ModalImageDelete/index.ts
@@ -1,1 +1,0 @@
-export { default as ModalImageDelete } from './ModalImageDelete'
