--- pro/src/components/ImageUploader/ButtonImageDelete/index.ts
@@ -1,1 +1,0 @@
-export { default as ButtonImageDelete } from './ButtonImageDelete'
