--- pro/src/components/ImageUploader/ButtonImageEdit/ModalImageEdit/ModalImageCrop/index.ts
@@ -1,1 +1,0 @@
-export { default as ModalImageCrop } from './ModalImageCrop'
