--- pro/src/components/ImageUploader/ButtonImageEdit/ModalImageEdit/ModalImageCrop/utils/index.ts
@@ -1,1 +1,0 @@
-export { default as getCropMaxDimension } from './getCropMaxDimension'
