--- pro/src/components/ImageUploader/ButtonImageEdit/ModalImageEdit/ModalImageUploadBrowser/index.ts
@@ -1,1 +1,0 @@
-export { default as ModalImageUploadBrowser } from './ModalImageUploadBrowser'
