--- pro/src/components/ImageUploader/ButtonImageEdit/ModalImageEdit/ModalImageUploadConfirm/index.ts
@@ -1,1 +1,0 @@
-export { default as ModalImageUploadConfirm } from './ModalImageUploadConfirm'
