--- pro/src/components/ImageUploader/ButtonImageEdit/ModalImageEdit/index.ts
@@ -1,1 +1,0 @@
-export { default as ModalImageEdit } from './ModalImageEdit'
