--- pro/src/components/ImageUploader/ButtonImageEdit/index.ts
@@ -1,5 +1,0 @@
-export { default as ButtonImageEdit } from './ButtonImageEdit'
-export type {
-  UploadImageValues as UploadImageValues,
-  CropParams as CropParams,
-} from './types'
