--- pro/src/components/ImageUploader/index.ts
@@ -1,2 +1,0 @@
-export { default as ImageUploader } from './ImageUploader'
-export type { UploadImageValues, CropParams } from './ButtonImageEdit'
